third_party/blink/web_tests/fast/forms/select/select-selectedIndex-multiple-expected.txt

CONSOLE WARNING: Unable to expand the option list to length 4294967295. The maximum allowed list length is 100000.
CONSOLE WARNING: Unable to expand the option list to length 100001. The maximum allowed list length is 100000.
one
two
1) setting length to a negative length
PASS mySelect.options.length is 2
PASS mySelect.selectedIndex is 0
2) setting length to a larger length
PASS mySelect.options.length is 5
PASS mySelect.selectedIndex is 0
3) setting length to a smaller length
PASS mySelect.options.length is 2
PASS mySelect.selectedIndex is 0
PASS mySelect.options.length is 1
PASS mySelect.selectedIndex is 0
4) setting length to the same length
PASS mySelect.options.length is 2
PASS mySelect.selectedIndex is 0
5) setting length to non-integer value: null
PASS mySelect.options.length is 0
PASS mySelect.selectedIndex is -1
6) setting length to non-integer value: undefined
PASS mySelect.options.length is 0
PASS mySelect.selectedIndex is -1
7) setting length to non-integer value: true
PASS mySelect.options.length is 1
PASS mySelect.selectedIndex is 0
8) setting length to non-integer value: false
PASS mySelect.options.length is 0
PASS mySelect.selectedIndex is -1
9) setting length to non-integer value: non-numeric string
PASS mySelect.options.length is 0
PASS mySelect.selectedIndex is -1
10) setting length to non-integer value: object
PASS mySelect.options.length is 0
PASS mySelect.selectedIndex is -1
11) setting length to non-integer value: negative infinity
PASS mySelect.options.length is 0
PASS mySelect.selectedIndex is -1
12) setting length to non-integer value: NaN
PASS mySelect.options.length is 0
PASS mySelect.selectedIndex is -1
13) setting length to non-integer value: positive infinity
PASS mySelect.options.length is 0
PASS mySelect.selectedIndex is -1
14) setting length to non-integer value: floating point number
PASS mySelect.options.length is 2
PASS mySelect.selectedIndex is 0
15) setting an element by index past the end of the current list
PASS mySelect.options.length is 11
PASS mySelect.selectedIndex is 0
16) setting an existing element by index
PASS mySelect.options.length is 11
PASS mySelect.selectedIndex is 0
17) trying to set an element that's not an option: null
PASS mySelect.options.length is 10
PASS mySelect.selectedIndex is 0
18) trying to set an element that's not an option: undefined
PASS mySelect.options.length is 10
PASS mySelect.selectedIndex is 0
19) trying to set an element that's not an option: select element
PASS mySelect.options[10] = mySelect; threw exception TypeError: Failed to set an indexed property [10] on 'HTMLOptionsCollection': parameter 2 is not of type 'HTMLOptionElement'..
PASS mySelect.options.length is 10
PASS mySelect.selectedIndex is 0
20) trying to set a option element using an invalid index: negative infinity
PASS mySelect.options.length is 10
PASS mySelect.selectedIndex is 0
21) trying to set a option element using an invalid index: NaN
PASS mySelect.options.length is 10
PASS mySelect.selectedIndex is 0
22) trying to set a option element using an invalid index: positive infinity
PASS mySelect.options.length is 10
PASS mySelect.selectedIndex is 0
23) setting length to a value greater than 100,000
PASS mySelect.options.length is 10
PASS mySelect.selectedIndex is 0


PASS successfullyParsed is true

TEST COMPLETE
